Restore python2 compatiblity and move tests from kernel.py to tests folder. Test not passing (yet!)
